refactor(api): move middlewares to `middlewares` package
